Circular Linked List Deletions
You will be implementing functions to delete nodes from a circular linkedlist.

Note that unlike previous times where we stores the head nodes of the linkedlist, we store the tail node of the linked list while using a circular linkedlist to make insertion operations more efficient.

You will have to copy the part for inserting nodes from the previous problem

The following is the structure of the circular linked list node :

struct LinkedListNode {
	int val;
    struct LinkedListNode* next;
};
typedef struct LinkedListNode LinkedListNode;
Additionally, here are the details of the functions that you are expected to implement :

LinkedListNode* deleteAtBeginning(LinkedListNode* tail)
/*
 * Deletes the node at the beginning of the linked list and
 * returns a pointer to the tail of the new linked list.
 * returns NULL if tail is NULL
 */

LinkedListNode* deleteAtEnd(LinkedListNode* tail)
/*
 * Deletes the node at the end of the linked list and
 * returns a pointer to the tail of the new list
 * returns NULL if no last node present
 */
Input
Not Needed

Output
Not Needed

Instructions
You are only allowed to edit the Body of the code
The Head and the Tail of the code must remain as it is
You are not responsible for taking any input/output
You must write the function with names exactly as given above
You must not create any global variables
Sample Input 0

10 10
Sample Output 0

correct