Circular Linked List Insertions
You will be implementing functions to insert nodes into a circular linkedlist.

Note that unlike previous times where we stores the head nodes of the linkedlist, we store the tail node of the linked list while using a circular linkedlist to make insertion operations more efficient.

The following is the structure of the linked list node :

struct LinkedListNode {
	int val;
    struct LinkedListNode* next;
};
typedef struct LinkedListNode LinkedListNode;
Here are the details of the functions that you are expected to implement :

LinkedListNode* insertAtBeginning(LinkedListNode* tail, int value)
/*
 * Inserts a node at the beginning of the circular linked list and
 * returns a pointer to the tail of the new list
 */

LinkedListNode* insertAtEnd(LinkedListNode* tail, int value)
/*
 * Inserts a node at the end of the circular linked list and
 * returns a pointer to the tail of the new list
 */
Input
Not Needed

Output
Not Needed

Instructions
You are only allowed to edit the Body of the code
The Head and the Tail of the code must remain as it is
You are not responsible for taking any input/output
You must write the function with names exactly as given above
You must not create any global variables
Sample Input 0

10 10
Sample Output 0

correct