Group of Jackals
The function accepts a positive integer array ‘array’ of size ’n’ as its argument. Implement the function to find the Jackal elements in the array and return their sum.

An element is a Jackal in the array if it is STRICTLY greater than all the elements to its right side. the rightmost element is always a Jackal.

Input Format

Complete the given Function

Constraints

The answer would lie in Integer range.

Output Format

Return the sum of the Jackal elements of the array

Sample Input 0

5
5 4 3 2 1
Sample Output 0

15