Matrix Addition
Just as 2D matrix, you can declare an array of as many dimensions as you want.

For example : 3D array : int arr3d[5][5][5], 4D array : int arr4d[5][5][5][5] and so on.

For this task, you are to write a program that adds two 3x3 matrices and prints the result

Sample Input 0

1 2 3
4 5 6
7 8 9
3 2 1
6 5 4
9 8 7
Sample Output 0

4 4 4
10 10 10
16 16 16