Linked Lists - Searching
Write a program to create a linked list with the given number of elements(follow the normal insertion process i.e., insert node at end evry time).If there are no elements given then the head of the linked list will be null. After completing the insertion of all you will be given an element you have to search for that element in the linked list. Display "yes" if it is found else "no".

Note: Intially the linked list will be empty

Input Format

First Line contains number of elements(n) to be inserted and followed the n+1 elements the (n+1)th element is the element to be searched

Constraints

0<=n<=100

1<=element<=10000

Output Format

"yes" or "no"

Sample Input 0

4
1
2
3
4
4
Sample Output 0

yes
Explanation 0

First 4 indicates the elements to be inserted followed by 4 elements to be inserted.         After insertion the list wil be 1->2->3->4->NULL

The last 4 is the element to be searched. Since 4 is present in the list we have printed      yes.