Height of a Binary Tree
Your task is to implement the following function :

int height(TreeNode*)
Height of a binary tree is defined as the maximum number of edges encountered on any path from the root node till a leaf node.

You will be working with the following structure :

struct TreeNode {
	int x;
    struct TreeNode* L;
    struct TreeNode* R;
}
You may only edit the BODY of the code, leaving the HEAD and the TAIL as it is.

Sample Input 0

7
4 2 1 3 6 7 5
Sample Output 0

2